Updated to Python3, cleaned up code, added custom workflows
Updated the code to be cleaner, better formatted, and compatible with
Python3.  Also, added the ability for limited use custom workflow input
with the `--CustomWorkflow` option, which will help in the cases where
we need to just do a quick test on a system that we don't normally use
and therefore dont' want to set up a config entry.